+++ src/fukinotou/parquet_loader.py
@@ -4,7 +4,7 @@ from typing import List, Type, TypeVar, Generic
 
 from pydantic import BaseModel, ValidationError
 
-from fukinotou.dataframe_exportable import DataframeExportable
+from fukinotou.abstraction.dataframe_exportable import DataframeExportable
 from fukinotou.exception.loading_exception import LoadingException
 
 T = TypeVar("T", bound=BaseModel)
